Points over finite field X_0^*(N), N squarre free
A pdf file, which contains Magma code for obtaining the modular forms for X_0^*(N), N square free, computing the number of points of X_0^*(N) over a finite field. Implementation of a method of Josep Gonzalez in order to obtain that the resudue at p prime with p nmid N of X_0^*(N) could have trivial automorphism group (where the automorphism group can only be a 2-group as is the case of X_0^*(N), with N square-free). An implementation given an elliptic curve $E$ with conductor dividing N if counting points over a finite field F we haver or not that $|X_0(N)^*(F)|\leq 2* |E(F)| $.